refactor: move packages under internal/ for encapsulation
