+++ internal/application/document_service_test.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 	"testing"
 
-	"ai-search-emulator/domain"
+	"ai-search-emulator/internal/domain"
 )
 
 func newDocumentServiceForTest() (*DocumentService, *mockIndexRepository, *mockDocumentRepository) {
